Current WIP
This SHOULD work to a degree.  Right now the logic is being added to the code in so that if a player has a natural 21 they automatically win unless the dealer also has a natural 21.  Also the ability to actually bet money needs to be added for it be as real as possible.